fixing broken php unit tests

+++ app/Http/Controllers/ContactsController.php
@@ -32,11 +32,19 @@ class ContactsController extends Controller
     public function update(Contact $contact)
     {
         
+        if(request()->user()->isNot($contact->user)) {
+            return response([], 403);
+        }
+        
         $contact->update($this->validateData());
     }
 
     public function destroy(Contact $contact)
     {
+        if(request()->user()->isNot($contact->user)) {
+            return response([], 403);
+        }
+        
         $contact->delete();
     }
